+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * Require `mfa` for releasing gem
 * Fix issues from new rubocop version `v1.24.0`
+* Remove deprecated `OnlyofficeFileHelper.get_filename`
 
 ## 0.5.0 (2020-11-28)